Added refactor changes and optimised pdf view.
