+++ app/views/report/resident/gainSummaryReport.scala.html
@@ -7,16 +7,12 @@
 
 @(answers: YourAnswersSummaryModel, gain: BigDecimal, taxYear: TaxYearModel)(implicit request: Request[_])
 
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml">
+<!DOCTYPE html>
+<html>
 
 <head>
     <meta name="viewport" content="width=device-width" />
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-    <link href="/template/assets/stylesheets/govuk-template.css" media="screen" rel="stylesheet" type="text/css">
-    <link href="/template/assets/stylesheets/govuk-template-print.css" media="print" rel="stylesheet" type="text/css">
-    <link href="/template/assets/stylesheets/fonts.css" media="all" rel="stylesheet" type="text/css">
-    <link rel="stylesheet" href="http://localhost:9032/assets/2.203.0/stylesheets/application.min.css">
     <link rel="stylesheet" type="text/css" href="/calculate-your-capital-gains/non-resident/assets/stylesheets/cgt.css">
     <title>@Messages("calc.resident.summary.title")</title>
 </head>
