+++ test/controllers/resident/ReportControllerTests/DeductionsSummaryActionSpec.scala
@@ -14,11 +14,13 @@
  * limitations under the License.
  */
 
-package views.resident.report
+package controllers.resident.ReportControllerTests
 
 import controllers.helpers.FakeRequestHelper
+import org.scalatest.mock.MockitoSugar
 import uk.gov.hmrc.play.test.{UnitSpec, WithFakeApplication}
 
-class DeductionsSummaryReportViewSpec extends UnitSpec with WithFakeApplication with FakeRequestHelper{
+class DeductionsSummaryActionSpec extends UnitSpec with WithFakeApplication with FakeRequestHelper with MockitoSugar {
 
 }
+
