+++ test/views/resident/report/DeductionsSummaryViewSpec.scala
@@ -14,13 +14,11 @@
  * limitations under the License.
  */
 
-package controllers.resident.ReportControllerTests
+package views.resident.report
 
 import controllers.helpers.FakeRequestHelper
-import org.scalatest.mock.MockitoSugar
 import uk.gov.hmrc.play.test.{UnitSpec, WithFakeApplication}
 
-class FinalSummaryReportActionSpec extends UnitSpec with WithFakeApplication with FakeRequestHelper with MockitoSugar{
+class DeductionsSummaryViewSpec extends UnitSpec with WithFakeApplication with FakeRequestHelper{
 
 }
-
