+++ test/views/resident/report/FinalSummaryViewSpec.scala
@@ -14,12 +14,11 @@
  * limitations under the License.
  */
 
-package controllers.resident.ReportControllerTests
+package views.resident.report
 
 import controllers.helpers.FakeRequestHelper
-import org.scalatest.mock.MockitoSugar
 import uk.gov.hmrc.play.test.{UnitSpec, WithFakeApplication}
 
-class GainSummaryReportActionSpec extends UnitSpec with WithFakeApplication with FakeRequestHelper with MockitoSugar{
+class FinalSummaryViewSpec extends UnitSpec with WithFakeApplication with FakeRequestHelper{
 
 }
